refactor: move converter logic to separate package
