+++ index.d.ts
@@ -1,2 +1,3 @@
 declare const ansiEscapesSerializer: import("pretty-format").NewPlugin;
+
 export = ansiEscapesSerializer;
