--- src/index.js
@@ -1,3 +1,0 @@
-const { ansiEscapesSerializer } = require("./ansiEscapesSerializer");
-
-module.exports = ansiEscapesSerializer;
