+++ tests/cursor-and-erase.test.js
